requirements.txt
async-generator==1.10
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
black==22.12.0
cachetools==5.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
cryptography==37.0.4
Deprecated==1.2.13
ffmpeg-python==0.2.0
filelock==3.9.0
flake8==6.0.0
Flask==2.2.2
Flask-Cors==3.0.10
future==0.18.3
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
h11==0.13.0
httplib2==0.20.4
huggingface-hub==0.12.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
limits==2.7.0
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
more-itertools==9.0.0
mypy-extensions==0.4.3
numpy==1.24.1
openai-whisper==20230124
outcome==1.2.0
packaging==21.3
pandas==1.5.3
pathspec==0.11.0
platformdirs==2.6.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==12.5.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.1.0
xlrd==2.0.1
